Adds truly scalable active indicators
The script will now be truly scalable in terms of showing what location and filter is currently selected and changing the color of that one to the parameter. Now all changes to the content only involve the html

I also replaced all for loops back to i for the counter since it was stupid to change them in the first place.